imp(game): moved out game to component file.
